Got basic pathfinder working
This is the first pass at a platformer pathfinder algorithm.

The basic idea is very similar to basic maze generation logic - with one
key difference: instead of generating a typical maze, this algorithm
generates a path through a space that coudl work for a 2D/2.5D platform
game (similar to Mario or Sonic).

This difference is important as it comes with one requirement: there must
be a single, unbroken, path from the left side of the space to the right
side.

There can be branches off the main path (indeed, that is something I want
to implement soon), but there has to be at least one path a player can take
from the "start" of the map to the "end".