Convert @farthest_right to an object, handle all undefined cases
So the pathfinding logic was a bit wonky - mostly for two reasons:

1. All `undefined` cases were not being handled properly
2. There was no good way to get out of a situation where the pathfinder
   had nowhere to go

This commit fixes both issues.

The biggest change is that `@farthest_right` is now an object storing both
the `col` and `row` of the further carved path to the right. This is important
because it gives the pathfinder a place to go bac kto if it gets stuck. This solves
one of the `undefined` cases and 2 from above.

Another fix is that if `@director` returns `undefined` `@carve_path()` is called again.
The idea here is that `@director` should not return `undefined`, but if/when it does
trying again is the best solution.

This surfaces an api issue: it needs to be made clear that `@director` should not ever
return `undefined`.